simple user agent service
