--- test/Services/ParserService.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-using Xunit;
-
-namespace Wangkanai.Detection.Services
-{
-    public class ParserService
-    {
-        [Fact]
-        public void Ctor_Null_ThrowsArgumentNullException()
-        {
-            Assert.Throws<ArgumentNullException>(() => new ParserService(null));
-        }
-    }
-}
